requirements.txt
apache-airflow==2.7.3
kafka-python==2.0.2
psycopg2-binary==2.9.9
requests==2.31.0
pandas==2.1.4
sqlalchemy==2.0.23